Added basic UI elements
Modified the MockPlantProvider class since it uses a substantial amount of logic. Using a lambda expression since it's super simple to implement. It's more advanced than what we've been using so far, but good to take a look at!

Added an ID field to all PlantModels, so we can sort easier and decided on a 1-based index to make it more like a real database.

The next task is to use the drop down as a way to select the "amount" of items, rather than the index of the item. Using it as the index was useful to test with, but it's definitely not what we want.